fix: separate utils from React components to fix Fast Refresh
Resolve `react-refresh/only-export-components` linting errors that caused
the CI check suite to fail.

Extracted shared constants like Framer Motion variants (`rowVariants`) and
helper functions (`getStatusBadge`) into a dedicated `utils.ts` file so
that `BatchJobTableRow.tsx` solely exports its React component. All
affected frontend tests run successfully.
